refactor(signalprocessing): Phase 6 Wave 2b - decompose Reconcile/reconcileEnriching/reconcileClassifying, split file
Decomposes the three highest-complexity functions in the SignalProcessing
controller:

- Reconcile: 23 -> 15 cyclomatic, via initializeNewSignalProcessing,
  dispatchPhase, and isDuplicateGenerationReconcile (the last extracted after
  a first pass left the dispatcher at 18).
- reconcileEnriching: 38 -> 3, via checkEnrichingIdempotencyGuard,
  warnIfUnsupportedTargetType, performK8sEnrichment,
  applyEnrichmentCustomLabels, buildEnrichmentMessage, finalizeEnrichment.
- reconcileClassifying: 36 -> 8, via checkClassifyingIdempotencyGuard,
  refreshClassifyingContext, failClassifyingPhase, evaluateSeverityOrFail,
  resolveSignalMode, buildClassificationMessage, finalizeClassification.

Splits signalprocessing_controller.go (1,336 lines post-decomposition) into
signalprocessing_controller.go (core dispatcher + reconcilePending + backoff
helpers), signalprocessing_enriching.go, signalprocessing_classifying.go,
signalprocessing_categorizing.go, and signalprocessing_audit.go. All files
now under the 700-line convention threshold.

Per GO-ANTIPATTERN-AUDIT-2026-07-01 Wave 2 (issue #1520, see §7j-2).
